[ch1365] Moving data fixture to dashboard props.
