+++ src/routes/AppRouter.js
@@ -4,7 +4,6 @@ import { createBrowserHistory } from "history"
 
 export const history = createBrowserHistory()
 
-const Home = lazy(() => import("components/Home/Home"))
 const Overview = lazy(() => import("components/Overview/Overview"))
 const Dashboard = lazy(() => import("components/Dashboard/Dashboard"))
 
@@ -15,8 +14,7 @@ const AppRouter = () => (
       <Switch>
         <Route exact path='/dashboard' component={Dashboard} />
         <Route exact path='/overview' component={Overview} />
-        <Route exact path='/home' component={Home} />
-        <Route exact path='/' component={Home} />
+        <Route exact path='/' component={Overview} />
       </Switch>
     </Suspense>
   </Router>
